refactor: 랜덤 로또 생성 로직을 LottoGenerator 로 이동
